Resolves bug 3109348 (OOM during MRCNN training with batch size = 1).

+++ tensorflow/core/kernels/data/experimental/map_and_batch_dataset_op.cc
@@ -549,9 +549,10 @@ class MapAndBatchDatasetOp : public UnaryDatasetOpKernel {
         auto busy = [this]() EXCLUSIVE_LOCKS_REQUIRED(*mu_) -> bool {
           int64 num_parallel_calls = num_parallel_calls_->value;
           return num_calls_ >= num_parallel_calls ||
-                 (batch_results_.size() > max_batch_results_ ||
-                  (batch_results_.size() == max_batch_results_ &&
-                   call_counter_ % dataset()->batch_size_ == 0));
+              (static_cast<int64>(batch_results_.size()) > max_batch_results_ ||
+              (static_cast<int64>(batch_results_.size()) ==
+	      max_batch_results_ &&
+              call_counter_ % dataset()->batch_size_ == 0));
         };
         while (true) {
           {
